Time: 44 ms (40.31%), Space: 19.5 MB (11.97%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -9,7 +9,7 @@ public:
         int ans=INT_MAX;
         while(l<=e)
         {
-            long long sm=0;
+            unsigned int sm=0;
             mid=l+(e-l)/2;
             for (int i = 0; i < piles.size(); i ++) 
             {
